requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
oracledb==2.0.1
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

